最終調整分追加

+++ Assets/Scripts/Core/GameEvents.cs
@@ -19,7 +19,7 @@ namespace BlockBlastGame
         public static event Action<List<ItemData>> OnSpaceshipBuild;
 
         // Calorie events
-        public static event Action<int> OnCalorieChanged; // totalCalories
+        public static event Action<long> OnCalorieChanged; // totalCalories
 
         // Wave events
         public static event Action<int, int> OnWaveStarted;            // waveIndex, totalWaves
@@ -49,7 +49,7 @@ namespace BlockBlastGame
         public static void TriggerStageClear() => OnStageClear?.Invoke();
         public static void TriggerSpaceshipBuild(List<ItemData> parts) => OnSpaceshipBuild?.Invoke(parts);
 
-        public static void TriggerCalorieChanged(int totalCalories) => OnCalorieChanged?.Invoke(totalCalories);
+        public static void TriggerCalorieChanged(long totalCalories) => OnCalorieChanged?.Invoke(totalCalories);
 
         public static void TriggerWaveStarted(int waveIndex, int totalWaves) => OnWaveStarted?.Invoke(waveIndex, totalWaves);
         public static void TriggerSurvivalTimerUpdate(float elapsed, float limit) => OnSurvivalTimerUpdate?.Invoke(elapsed, limit);
